% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insect_counts}
\alias{insect_counts}
\title{Insect sweep net data}
\format{A data.frame with 5 variables: \code{date}, \code{watershed},
\code{rep}, \code{taxa}, and \code{count}}
\usage{
insect_counts
}
\description{
Counts of all insects
}
\examples{
library(dplyr)
with_guild <- insect_count \%>\%
  left_join(insect_guild, by="taxa")
}
\seealso{
\code{\link{insect_guilds}}
}
\keyword{datasets}

