% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insect_guilds}
\alias{insect_guilds}
\title{Insect guilds}
\format{A data.frame with two variables: \code{taxa} and
\code{guild}}
\usage{
insect_guilds
}
\description{
Insect guilds
}
\examples{
library(dplyr)
with_guild <- insect_count \%>\%
  left_join(insect_guilds, by="taxa")
}
\seealso{
\code{\link{insect_counts}}
}
\keyword{datasets}

